refactor(ui): layout component for pages
