+++ ui/src/components/Sidebar/SidebarRow.tsx
@@ -10,7 +10,10 @@ interface Props {
 
 export const SidebarRow: React.FC<Props> = props => {
   return (
-    <div className={styles.container_row} onClick={props.onClick}>
+    <div
+      className={styles.container_row}
+      onClick={() => props.onClick && props.onClick()}
+    >
       <div className={styles.row_content}>
         <div className={styles.icondesign}>
           <img src={props.icon} alt="icon" />
