+++ ui/src/components/TopBar/styles.module.scss
@@ -1,3 +1,5 @@
+@import '../../index';
+
 .logoContainer {
   padding: 1rem;
   width: 60px;
@@ -36,7 +38,6 @@
   color: var(--primary-text);
   margin-left: 1vh;
 }
-@import '../../index';
 
 @media only screen and (min-width: $sm) {
   .logoContainer {
