--- ui/src/pages/Profile/styles.module.scss
@@ -1,23 +1,0 @@
-.container {
-  min-height: 100vh;
-  width: 100vw;
-  background-color: #f4f5f7;
-  display: flex;
-}
-.NavContainer {
-  width: 18%;
-  background-color: pink;
-}
-.mainContent {
-  width: 82%;
-}
-
-@import '../../index';
-@media only screen and (max-width: $sm) {
-  .NavContainer {
-    display: none;
-  }
-  .mainContent {
-    width: 100%;
-  }
-}
